Various fixes with indexing
Fixed problem with thread workers not getting session settings. The session dictionary is now written to temp/session.json for the workers to read from during their init.
Fixed problem with thread method resulting in inconsistent results due to the actual directory not being checked for being a 'photo' directory.

utils/infrequents: added function to_structure(), takes in a dict to create a directory structure which reflects the dictionary structure.